+++ packages/devtools_extensions/bin/_build_and_copy.dart
